gitsphere-backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.4.2
python-dotenv==1.0.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
PyJWT=2.10.1